Minor bugfixes. Fixed another series of warnings.
git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@4539 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ base/gpe-conf/sleep/callbacks.c
@@ -16,6 +16,7 @@
 #include "conf.h"
 #include "confGUI.h"
 #include "../applets.h"
+#include "../ipaqscreen/brightness.h"
 
 extern GtkWidget *sleep_idle_spin;
 extern GtkWidget *dim_spin;
@@ -264,7 +265,7 @@ on_dim_scale_focus_out_event            (GtkWidget       *widget,
                                         GdkEventFocus   *event,
                                         gpointer         user_data)
 {
-  set_brightness(user_data);
+  set_brightness((int)user_data);
   return FALSE;
 }
